+++ endpoint-viz/location.py
@@ -1,0 +1,5 @@
+class Location:
+    def __init__(self, r, c, path):
+        self.r    = r
+        self.c    = c
+        self.path = path
